# Improve `ServiceManager#has()` performance