@boesing 2021-07-24
@Ocramius I've added phpbench in #94 so we can rebase this branch against 3.7.x afterwards.