@boesing 2021-07-24
## Benchmark on my MBP 2019 2,8 GHz i7

### Before

```
PHPBench @git_tag@ running benchmarks...
with configuration file: /Users/max/git/laminas/laminas-servicemanager/phpbench.json
with PHP version 7.4.21, xdebug ✔, opcache ❌

\LaminasBench\ServiceManager\HasNewServicesBench

    benchHasFactory1........................R1 I1 [μ Mo]/r: 3.032 3.073 (μs) [μSD μRSD]/r: 0.068μs 2.23%
    benchHasInvokable1......................R1 I9 [μ Mo]/r: 3.813 3.796 (μs) [μSD μRSD]/r: 0.081μs 2.14%
    benchHasService1........................R2 I9 [μ Mo]/r: 3.093 3.054 (μs) [μSD μRSD]/r: 0.073μs 2.36%
    benchFetchFactoryAlias1.................R1 I6 [μ Mo]/r: 3.736 3.788 (μs) [μSD μRSD]/r: 0.083μs 2.22%
    benchHasRecursiveFactoryAlias1..........R3 I6 [μ Mo]/r: 3.839 3.861 (μs) [μSD μRSD]/r: 0.065μs 1.70%
    benchFetchRecursiveFactoryAlias2........R2 I8 [μ Mo]/r: 3.963 3.986 (μs) [μSD μRSD]/r: 0.102μs 2.58%
    benchFetchAbstractFactoryFoo............R1 I0 [μ Mo]/r: 4.821 4.778 (μs) [μSD μRSD]/r: 0.087μs 1.81%

(best [mean mode] worst) = 2.923 [3.757 3.762] 3.128 (μs)
⅀T: 262.978μs μSD/r 0.080μs μRSD/r: 2.148%
Subjects: 7, Assertions: 0, Warnings: 0, Errors: 0, Failures: 0
```

```
→ php74 vendor/bin/phpbench  run benchmarks/HasCachedServicesBench.php 
PHPBench @git_tag@ running benchmarks...
with configuration file: /Users/max/git/laminas/laminas-servicemanager/phpbench.json
with PHP version 7.4.21, xdebug ✔, opcache ❌

\LaminasBench\ServiceManager\HasCachedServicesBench

    benchHasFactory1........................R1 I19 [μ Mo]/r: 2.901 2.823 (μs) [μSD μRSD]/r: 0.093μs 3.20%
    benchHasInvokable1......................R1 I12 [μ Mo]/r: 3.612 3.574 (μs) [μSD μRSD]/r: 0.090μs 2.48%
    benchHasService1........................R2 I17 [μ Mo]/r: 2.999 3.022 (μs) [μSD μRSD]/r: 0.078μs 2.60%
    benchHasAlias1..........................R1 I10 [μ Mo]/r: 3.130 3.160 (μs) [μSD μRSD]/r: 0.069μs 2.22%
    benchHasRecursiveAlias1.................R2 I17 [μ Mo]/r: 3.119 3.092 (μs) [μSD μRSD]/r: 0.074μs 2.38%
    benchHasRecursiveAlias2.................R1 I3 [μ Mo]/r: 3.290 3.251 (μs) [μSD μRSD]/r: 0.090μs 2.72%
    benchHasAbstractFactoryService..........R2 I18 [μ Mo]/r: 5.064 5.046 (μs) [μSD μRSD]/r: 0.097μs 1.92%

(best [mean mode] worst) = 2.785 [3.445 3.424] 3.040 (μs)
⅀T: 482.300μs μSD/r 0.084μs μRSD/r: 2.504%
Subjects: 7, Assertions: 0, Warnings: 0, Errors: 0, Failures: 0
```

## After
```
→ php74 vendor/bin/phpbench  run benchmarks/HasCachedServicesBench.php 
PHPBench @git_tag@ running benchmarks...
with configuration file: /Users/max/git/laminas/laminas-servicemanager/phpbench.json
with PHP version 7.4.21, xdebug ✔, opcache ❌

\LaminasBench\ServiceManager\HasCachedServicesBench

    benchHasFactory1........................R1 I4 [μ Mo]/r: 0.571 0.580 (μs) [μSD μRSD]/r: 0.014μs 2.49%
    benchHasInvokable1......................R1 I8 [μ Mo]/r: 0.889 0.881 (μs) [μSD μRSD]/r: 0.021μs 2.34%
    benchHasService1........................R1 I1 [μ Mo]/r: 0.571 0.570 (μs) [μSD μRSD]/r: 0.014μs 2.50%
    benchHasAlias1..........................R1 I7 [μ Mo]/r: 0.595 0.594 (μs) [μSD μRSD]/r: 0.012μs 2.02%
    benchHasRecursiveAlias1.................R1 I1 [μ Mo]/r: 0.606 0.602 (μs) [μSD μRSD]/r: 0.012μs 2.01%
    benchHasRecursiveAlias2.................R2 I17 [μ Mo]/r: 0.608 0.612 (μs) [μSD μRSD]/r: 0.011μs 1.87%
    benchHasAbstractFactoryService..........R1 I4 [μ Mo]/r: 1.171 1.181 (μs) [μSD μRSD]/r: 0.025μs 2.10%

(best [mean mode] worst) = 0.545 [0.716 0.717] 0.590 (μs)
⅀T: 100.238μs μSD/r 0.016μs μRSD/r: 2.193%
Subjects: 7, Assertions: 0, Warnings: 0, Errors: 0, Failures: 0
```

```
→ php74 vendor/bin/phpbench  run benchmarks/HasNewServicesBench.php 
PHPBench @git_tag@ running benchmarks...
with configuration file: /Users/max/git/laminas/laminas-servicemanager/phpbench.json
with PHP version 7.4.21, xdebug ✔, opcache ❌

\LaminasBench\ServiceManager\HasNewServicesBench

    benchHasFactory1........................R2 I8 [μ Mo]/r: 0.621 0.619 (μs) [μSD μRSD]/r: 0.011μs 1.77%
    benchHasInvokable1......................R1 I9 [μ Mo]/r: 0.954 0.934 (μs) [μSD μRSD]/r: 0.024μs 2.52%
    benchHasService1........................R1 I0 [μ Mo]/r: 0.600 0.592 (μs) [μSD μRSD]/r: 0.013μs 2.23%
    benchFetchFactoryAlias1.................R5 I9 [μ Mo]/r: 0.950 0.981 (μs) [μSD μRSD]/r: 0.033μs 3.45%
    benchHasRecursiveFactoryAlias1..........R2 I8 [μ Mo]/r: 0.942 0.924 (μs) [μSD μRSD]/r: 0.026μs 2.80%
    benchFetchRecursiveFactoryAlias2........R1 I3 [μ Mo]/r: 0.995 1.013 (μs) [μSD μRSD]/r: 0.029μs 2.94%
    benchFetchAbstractFactoryFoo............R2 I6 [μ Mo]/r: 1.107 1.114 (μs) [μSD μRSD]/r: 0.020μs 1.83%

(best [mean mode] worst) = 0.585 [0.881 0.882] 0.628 (μs)
⅀T: 61.683μs μSD/r 0.022μs μRSD/r: 2.506%
Subjects: 7, Assertions: 0, Warnings: 0, Errors: 0, Failures: 0
```