@boesing 2021-07-24
@koenkivits This is awesome. I've double checked your implementation locally and also tried avoiding recursive method calls with using `do-while`.
When using `do-while`, I get some better results while losing readability.

```php
/**
     * Check if a static service or factory exists for the given name.
     */
    private function staticServiceOrFactoryCanCreate(string $name): bool
    {
        $resolvedName = $name;
        do {
            if (isset($this->services[$resolvedName]) || isset($this->factories[$resolvedName])) {
                return true;
            }

            $previousName = $resolvedName;
        } while (($resolvedName = $this->aliases[$resolvedName] ?? $resolvedName) !== $previousName);

        return false;
    }

    /**
     * Check if an abstract factory exists that can create a service for the given name.
     */
    private function abstractFactoryCanCreate(string $name): bool
    {
        $resolvedName = $name;

        do {
            foreach ($this->abstractFactories as $abstractFactory) {
                if ($abstractFactory->canCreate($this->creationContext, $resolvedName)) {
                    return true;
                }
            }

            $previousName = $resolvedName;
        } while (($resolvedName = $this->aliases[$resolvedName] ?? $resolvedName) !== $previousName);

        return false;
    }
```
----

**\LaminasBench\ServiceManager\HasCachedServicesBench**

(best [mean mode] worst) = 0.552 [0.683 0.684] 0.581 (μs)
----
**\LaminasBench\ServiceManager\HasNewServicesBench**

(best [mean mode] worst) = 0.566 [0.747 0.746] 0.615 (μs)

----

Thoughts?