@boesing 2021-07-24
> Oh wow! Nice work. @boesing did you happen to see how many/if any opacaching made a difference during benchmarking?
Sadly no, I am not that familiar with all that opcache stuff yet as I mainly work with PHP 7.3 in my daily work.