@Ocramius 2021-07-24
I'll do a local run with:
 * PHP 8
 * OpCache 
 * disabled CPU turbo boosting (it breaks benchmark results massively otherwise 😬)

Then I'll most likely merge+release, and add the results as merge commit messages 👍

Fantastic work, @koenkivits and @boesing!