@koenkivits 2021-07-24
I wasn't able to comment since I posted this, but awesome to see this being followed up on so quickly. Thanks @boesing and @Ocramius!

@boesing thanks for checking and adding the benchmark! Regarding the `do-while`: I'm not against sacrificing a little bit of readability for better performance in cases like this, although I'm not sure if the difference in performance is enough in that case. It's hard for me to make that call as an outsider, though. :)