@Ocramius 2021-07-24
> I'm not sure if the difference in performance is enough in that case

`laminas/laminas-servicemanager` being at the very core of most of `laminas/*` kinda makes it worth it.